Half a fix for float errors on walls

+++ src/game/map_rendering.c
@@ -6,7 +6,7 @@
 /*   By: pdruart <pdruart@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/02/24 15:44:11 by pdruart       #+#    #+#                 */
-/*   Updated: 2022/03/13 18:09:18 by rdrazsky      ########   odam.nl         */
+/*   Updated: 2022/03/16 17:51:13 by rdrazsky      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -45,8 +45,8 @@ static void	static_draw_line(
 	{
 		c = ml_color_at(texture, texture->width
 				* (data->rays[i].pos_on_wall), j * texel_step);
-		ml_draw_pixel(data->screen, i, data->player_ud_angle
-			+ j + (HEIGHT - height) / 2, ml_color_darken(c, darkness_mod));
+		ml_draw_pixel(data->screen, i, (int)data->player_ud_angle
+			+ j + (int)(HEIGHT - height) / 2, ml_color_darken(c, darkness_mod));
 		j++;
 	}
 }
